refactor(node): Extract PeerNode runtime and handshake lifecycle
Split the large PeerNode internals into dedicated runtime and handshake lifecycle helpers, preserving existing behavior while reducing class dependency pressure.

Also adapt noderef validation/tests, add runtime and lifecycle unit coverage, and document the new helper classes with doclint-clean Javadocs.
