+++ src/test/java/network/crypta/node/PeerNodeAddressManagerTest.java
@@ -15,8 +15,8 @@ import static org.mockito.Mockito.when;
 import java.lang.reflect.Field;
 import java.net.InetAddress;
 import java.net.UnknownHostException;
-import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.CopyOnWriteArrayList;
 import java.util.concurrent.TimeUnit;
 import network.crypta.io.comm.FreenetInetAddress;
 import network.crypta.io.comm.Peer;
@@ -71,7 +71,7 @@ class PeerNodeAddressManagerTest {
       throws ReflectiveOperationException {
     // Arrange
     PeerNodeAddressManager manager = new PeerNodeAddressManager(peer);
-    peer.nominalPeer = new ArrayList<>();
+    peer.nominalPeer = new CopyOnWriteArrayList<>();
     setLastAttemptedHandshakeIpUpdateTime(
         manager, System.currentTimeMillis() - TimeUnit.MINUTES.toMillis(1));
 
@@ -87,7 +87,7 @@ class PeerNodeAddressManagerTest {
       throws ReflectiveOperationException {
     // Arrange
     PeerNodeAddressManager manager = new PeerNodeAddressManager(peer);
-    peer.nominalPeer = new ArrayList<>();
+    peer.nominalPeer = new CopyOnWriteArrayList<>();
 
     // Act
     manager.maybeUpdateHandshakeIPs(true);
@@ -102,7 +102,7 @@ class PeerNodeAddressManagerTest {
       throws UnknownHostException {
     // Arrange
     PeerNodeAddressManager manager = new PeerNodeAddressManager(peer);
-    peer.nominalPeer = new ArrayList<>();
+    peer.nominalPeer = new CopyOnWriteArrayList<>();
     Peer detectedPeer = new Peer(inetAddress(203, 0, 113, 10), 1234);
     when(peer.getPeer()).thenReturn(detectedPeer);
 
@@ -137,7 +137,8 @@ class PeerNodeAddressManagerTest {
     Peer peerMatchingNode = new Peer(matchingAddress, 2222);
     Peer duplicatePeer = new Peer(matchingAddress, 2222);
 
-    peer.nominalPeer = new ArrayList<>(List.of(detectedDuplicate, peerMatchingNode, duplicatePeer));
+    peer.nominalPeer =
+        new CopyOnWriteArrayList<>(List.of(detectedDuplicate, peerMatchingNode, duplicatePeer));
 
     FreenetInetAddress localhost = new FreenetInetAddress(inetAddress(127, 0, 0, 1));
     when(node.network().freenetLocalhostAddress()).thenReturn(localhost);
@@ -173,7 +174,7 @@ class PeerNodeAddressManagerTest {
       throws ReflectiveOperationException {
     // Arrange
     PeerNodeAddressManager manager = new PeerNodeAddressManager(peer);
-    peer.nominalPeer = new ArrayList<>();
+    peer.nominalPeer = new CopyOnWriteArrayList<>();
     long before = System.currentTimeMillis();
 
     // Act
@@ -343,7 +344,7 @@ class PeerNodeAddressManagerTest {
     InetAddress address = inetAddress(198, 51, 100, 9);
     Peer nominal = new Peer(address, 2468);
     PeerNode peerNode = Mockito.mock(PeerNode.class, Mockito.CALLS_REAL_METHODS);
-    peerNode.nominalPeer = new ArrayList<>(List.of(nominal));
+    peerNode.nominalPeer = new CopyOnWriteArrayList<>(List.of(nominal));
     Mockito.doReturn(null).when(peerNode).getPeer();
 
     FreenetInetAddress matchAddress = new FreenetInetAddress(address);
@@ -361,7 +362,7 @@ class PeerNodeAddressManagerTest {
     InetAddress address = inetAddress(198, 51, 100, 10);
     Peer nominal = new Peer(address, 2468);
     PeerNode peerNode = Mockito.mock(PeerNode.class, Mockito.CALLS_REAL_METHODS);
-    peerNode.nominalPeer = new ArrayList<>(List.of(nominal));
+    peerNode.nominalPeer = new CopyOnWriteArrayList<>(List.of(nominal));
     Mockito.doReturn(null).when(peerNode).getPeer();
 
     FreenetInetAddress otherAddress = new FreenetInetAddress(inetAddress(203, 0, 113, 99));
